refactor: enrichment resolver 체인을 enrichment_network로 분리 (P2-A batch 2)
docs/refactoring-plan-2026-07.md §1.4 2단계에 따라 resolver/og-image 함수를
파사드(common/enrichment.py)에서 common/enrichment_network.py로 이동.

- 이동: _resolve_google_news_url, _resolve_google_news_url_inner,
  _resolve_via_gnewsdecoder, _fetch_og_image
- 부수 이동(순환 import 방지): _VERIFY_SSL, _get_verify_ssl
  (get_verify_ssl config import도 network로; 이동한 resolver가 호출)
- 파사드는 전 심볼 재-export(__all__ 불변 12), facade는 requests import 유지
  (fetch_page_metadata가 batch 4까지 사용)

@patch 재배치(84건, Correction D). 원칙: 심볼이 아니라 "호출자"가 이동해야 network로.
- _resolve_via_gnewsdecoder 17, _decode_google_news_base64 18 → 전부 network
  (유일 호출자 _inner이 이번에 이동 → batch 1에선 facade였던 것을 이번에 retarget)
- is_private_url 14 network / 6 facade (호출자별 split; 순수 함수라 HTTP가드가
  못 잡는 false-green을 리뷰어가 mock.called로 실증 검증)
- requests 35 network / 29 facade (requests는 공유 모듈이라 게이트엔 무영향,
  batch 4 facade import 제거 대비 정위치화)
- _resolve_google_news_url 8, _fetch_og_image 8 → 전부 facade 유지
  (호출자 fetch_*/fetch_page_metadata가 파사드 잔류; 6/2 split은 batch 3에서 완성)

facade 832→605줄, enrichment_network 404→642줄.
게이트: import OK(무순환), 715 passed, ruff clean, basedpyright 0 errors.
독립 코드리뷰 APPROVE(호출그래프+REPL mock.called 실증, 로직 변경 0 순수 이동).
