+++ tests/test_enrichment.py
@@ -322,7 +322,7 @@ class TestResolveGoogleNewsUrl:
         result = _resolve_google_news_url("https://reuters.com/markets/bitcoin")
         assert result == "https://reuters.com/markets/bitcoin"
 
-    @patch("common.enrichment._decode_google_news_base64")
+    @patch("common.enrichment_network._decode_google_news_base64")
     def test_base64_decode_success_skips_http(self, mock_decode):
         """If base64 decode succeeds, no network calls are made."""
         mock_decode.return_value = "https://real-article.com/news/123"
@@ -330,9 +330,9 @@ class TestResolveGoogleNewsUrl:
         assert result == "https://real-article.com/news/123"
         mock_decode.assert_called_once()
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder")
-    @patch("common.enrichment._decode_google_news_base64")
-    @patch("common.enrichment.requests.head")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder")
+    @patch("common.enrichment_network._decode_google_news_base64")
+    @patch("common.enrichment_network.requests.head")
     def test_http_head_redirect_used_when_base64_fails(self, mock_head, mock_decode, mock_gnews):
         """When base64 fails, follows HEAD redirect."""
         mock_gnews.return_value = ""
@@ -344,10 +344,10 @@ class TestResolveGoogleNewsUrl:
         result = _resolve_google_news_url("https://news.google.com/rss/articles/CBMiXXX")
         assert result == "https://real-site.com/article"
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder")
-    @patch("common.enrichment._decode_google_news_base64")
-    @patch("common.enrichment.requests.head")
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder")
+    @patch("common.enrichment_network._decode_google_news_base64")
+    @patch("common.enrichment_network.requests.head")
+    @patch("common.enrichment_network.requests.get")
     def test_get_fallback_when_head_stays_on_google(self, mock_get, mock_head, mock_decode, mock_gnews):
         """When HEAD stays on google, tries GET."""
         mock_gnews.return_value = ""
@@ -363,10 +363,10 @@ class TestResolveGoogleNewsUrl:
         result = _resolve_google_news_url("https://news.google.com/rss/articles/CBMiXXX")
         assert result == "https://real-site.com/article"
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder")
-    @patch("common.enrichment._decode_google_news_base64")
-    @patch("common.enrichment.requests.head")
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder")
+    @patch("common.enrichment_network._decode_google_news_base64")
+    @patch("common.enrichment_network.requests.head")
+    @patch("common.enrichment_network.requests.get")
     def test_network_exception_returns_empty(self, mock_get, mock_head, mock_decode, mock_gnews):
         """Network errors should be swallowed and return empty."""
         import requests as req_mod
@@ -893,10 +893,10 @@ class TestResolveGoogleNewsUrlHtmlBranch:
         # for the same URL do not bleed through as cache hits.
         _enrichment_mod._gnews_url_cache.clear()
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder")
-    @patch("common.enrichment._decode_google_news_base64")
-    @patch("common.enrichment.requests.head")
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder")
+    @patch("common.enrichment_network._decode_google_news_base64")
+    @patch("common.enrichment_network.requests.head")
+    @patch("common.enrichment_network.requests.get")
     def test_canonical_link_extracted_from_html(self, mock_get, mock_head, mock_decode, mock_gnews):
         """If GET stays on Google but HTML has canonical link, use it."""
         mock_gnews.return_value = ""
@@ -912,10 +912,10 @@ class TestResolveGoogleNewsUrlHtmlBranch:
         result = _resolve_google_news_url("https://news.google.com/rss/articles/CBMiHTML1")
         assert result == "https://real-article.com/page"
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder")
-    @patch("common.enrichment._decode_google_news_base64")
-    @patch("common.enrichment.requests.head")
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder")
+    @patch("common.enrichment_network._decode_google_news_base64")
+    @patch("common.enrichment_network.requests.head")
+    @patch("common.enrichment_network.requests.get")
     def test_og_url_extracted_from_html(self, mock_get, mock_head, mock_decode, mock_gnews):
         """og:url meta tag is used as fallback."""
         mock_gnews.return_value = ""
@@ -931,10 +931,10 @@ class TestResolveGoogleNewsUrlHtmlBranch:
         result = _resolve_google_news_url("https://news.google.com/rss/articles/CBMiHTML2")
         assert result == "https://real-article.com/og"
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder")
-    @patch("common.enrichment._decode_google_news_base64")
-    @patch("common.enrichment.requests.head")
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder")
+    @patch("common.enrichment_network._decode_google_news_base64")
+    @patch("common.enrichment_network.requests.head")
+    @patch("common.enrichment_network.requests.get")
     def test_get_exception_returns_empty(self, mock_get, mock_head, mock_decode, mock_gnews):
         """If GET raises an exception, return empty string."""
         import requests as req_mod
@@ -967,7 +967,7 @@ class TestFetchOgImage:
         _fetch_og_image = self._import_func()
         assert _fetch_og_image("") == ""
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_og_image_extracted(self, mock_get):
         from common.enrichment import _fetch_og_image
 
@@ -978,7 +978,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == "https://cdn.example.com/photo.jpg"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_twitter_image_fallback(self, mock_get):
         from common.enrichment import _fetch_og_image
 
@@ -989,7 +989,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == "https://cdn.example.com/tw.jpg"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_http_error_returns_empty(self, mock_get):
         import requests as req_mod
 
@@ -1001,7 +1001,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == ""
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_connection_error_returns_empty(self, mock_get):
         import requests as req_mod
 
@@ -1011,7 +1011,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == ""
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_invalid_image_url_skipped(self, mock_get):
         from common.enrichment import _fetch_og_image
 
@@ -1023,7 +1023,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == ""
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_no_og_image_returns_empty(self, mock_get):
         from common.enrichment import _fetch_og_image
 
@@ -1034,7 +1034,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == ""
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_og_image_secure_url_used(self, mock_get):
         """og:image:secure_url should be picked when og:image is absent."""
         from common.enrichment import _fetch_og_image
@@ -1046,7 +1046,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == "https://cdn.example.com/secure.jpg"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_article_image_used(self, mock_get):
         """article:image (OpenGraph article namespace) should be picked up."""
         from common.enrichment import _fetch_og_image
@@ -1058,7 +1058,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == "https://cdn.example.com/article.jpg"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_itemprop_image_used(self, mock_get):
         """Schema.org itemprop=image should be used as a late fallback."""
         from common.enrichment import _fetch_og_image
@@ -1070,7 +1070,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == "https://cdn.example.com/schema.jpg"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_link_image_src_used(self, mock_get):
         """<link rel="image_src"> legacy fallback should be picked up last."""
         from common.enrichment import _fetch_og_image
@@ -1082,7 +1082,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == "https://cdn.example.com/link-src.jpg"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_priority_og_over_twitter(self, mock_get):
         """og:image must win over twitter:image when both are present."""
         from common.enrichment import _fetch_og_image
@@ -1097,7 +1097,7 @@ class TestFetchOgImage:
         result = _fetch_og_image("https://example.com/article")
         assert result == "https://cdn.example.com/og.jpg"
 
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.requests.get")
     def test_logo_url_rejected_falls_through(self, mock_get):
         """A logo-ish og:image should be skipped in favor of a later valid URL."""
         from common.enrichment import _fetch_og_image
@@ -2804,8 +2804,8 @@ class TestResolveGoogleNewsReadUrl:
     def setup_method(self):
         _enrich_mod._gnews_url_cache.clear()
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64")
     def test_read_url_normalized_and_base64_retried(self, mock_decode, _mock_gnews):  # noqa: PT019
         # First call (original /read/ URL) returns "", second call (/rss/articles/) returns a URL
         mock_decode.side_effect = ["", "https://real-article.com/news/read-path"]
@@ -2824,8 +2824,8 @@ class TestResolveGoogleNewsInnerGnewsDecoderSuccess:
     def setup_method(self):
         _enrich_mod._gnews_url_cache.clear()
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="https://decoded.com/article")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="https://decoded.com/article")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
     def test_gnewsdecoder_result_returned_directly(self, _mock_b64, _mock_gnews):  # noqa: PT019
         result = _enrich_mod._resolve_google_news_url_inner("https://news.google.com/rss/articles/CBMiGNEWS", timeout=1)
         assert result == "https://decoded.com/article"
@@ -2841,9 +2841,9 @@ class TestResolveGoogleNewsInnerSSRFBlock:
     def setup_method(self):
         _enrich_mod._gnews_url_cache.clear()
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
-    @patch("common.enrichment.is_private_url", return_value=True)
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network.is_private_url", return_value=True)
     def test_ssrf_blocked_before_head_request_returns_empty(self, _mock_priv, _mock_b64, _mock_gnews):  # noqa: PT019
         result = _enrich_mod._resolve_google_news_url_inner("https://news.google.com/rss/articles/CBMiSSRF", timeout=1)
         assert result == ""
@@ -2859,10 +2859,10 @@ class TestResolveGoogleNewsInnerRedirectHops:
     def setup_method(self):
         _enrich_mod._gnews_url_cache.clear()
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
-    @patch("common.enrichment.is_private_url", return_value=False)
-    @patch("common.enrichment.requests.head")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.head")
     def test_redirect_to_non_google_domain_returned(self, mock_head, _mock_priv, _mock_b64, _mock_gnews):  # noqa: PT019
         mock_resp = MagicMock()
         mock_resp.status_code = 301
@@ -2871,11 +2871,11 @@ class TestResolveGoogleNewsInnerRedirectHops:
         result = _enrich_mod._resolve_google_news_url_inner("https://news.google.com/rss/articles/CBMiHOP", timeout=1)
         assert result == "https://real-news.com/article/123"
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
-    @patch("common.enrichment.is_private_url", return_value=False)
-    @patch("common.enrichment.requests.head")
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.head")
+    @patch("common.enrichment_network.requests.get")
     def test_empty_location_header_breaks_loop(self, mock_get, mock_head, _mock_priv, _mock_b64, _mock_gnews):  # noqa: PT019
         mock_resp = MagicMock()
         mock_resp.status_code = 301
@@ -2891,10 +2891,10 @@ class TestResolveGoogleNewsInnerRedirectHops:
         result = _enrich_mod._resolve_google_news_url_inner("https://news.google.com/rss/articles/CBMiEMPTY", timeout=1)
         assert result == ""
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
-    @patch("common.enrichment.is_private_url", return_value=False)
-    @patch("common.enrichment.requests.head")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.head")
     def test_relative_redirect_resolved_to_absolute(self, mock_head, _mock_priv, _mock_b64, _mock_gnews):  # noqa: PT019
         # First hop: relative redirect
         hop1 = MagicMock()
@@ -2908,10 +2908,10 @@ class TestResolveGoogleNewsInnerRedirectHops:
         result = _enrich_mod._resolve_google_news_url_inner("https://news.google.com/rss/articles/CBMiREL", timeout=1)
         assert "real-news.com" in result
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
-    @patch("common.enrichment.is_private_url", return_value=False)
-    @patch("common.enrichment.requests.head")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.head")
     def test_redirect_hop_to_private_ip_returns_empty(self, mock_head, _mock_priv, _mock_b64, _mock_gnews):  # noqa: PT019
         # Simulate redirect to a private IP (SSRF via redirect hop)
         mock_resp = MagicMock()
@@ -2919,7 +2919,7 @@ class TestResolveGoogleNewsInnerRedirectHops:
         mock_resp.headers = {"Location": "http://192.168.1.1/secret"}
         mock_head.return_value = mock_resp
         # Override is_private_url: allow google URL but block the redirected one
-        with patch("common.enrichment.is_private_url", side_effect=lambda u: "192.168" in u):
+        with patch("common.enrichment_network.is_private_url", side_effect=lambda u: "192.168" in u):
             result = _enrich_mod._resolve_google_news_url_inner(
                 "https://news.google.com/rss/articles/CBMiPRIV", timeout=1
             )
@@ -2933,16 +2933,16 @@ class TestResolveGoogleNewsInnerHeadFinalUrlSSRF:
     def setup_method(self):
         _enrich_mod._gnews_url_cache.clear()
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
-    @patch("common.enrichment.requests.head")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network.requests.head")
     def test_final_url_private_ip_returns_empty(self, mock_head, _mock_b64, _mock_gnews):  # noqa: PT019
         mock_resp = MagicMock()
         mock_resp.status_code = 200
         mock_resp.url = "http://10.0.0.1/internal"
         mock_resp.headers = {}
         mock_head.return_value = mock_resp
-        with patch("common.enrichment.is_private_url", side_effect=lambda u: "10.0.0" in u):
+        with patch("common.enrichment_network.is_private_url", side_effect=lambda u: "10.0.0" in u):
             result = _enrich_mod._resolve_google_news_url_inner(
                 "https://news.google.com/rss/articles/CBMiFINAL", timeout=1
             )
@@ -2959,9 +2959,9 @@ class TestResolveGoogleNewsInnerGetSSRF:
     def setup_method(self):
         _enrich_mod._gnews_url_cache.clear()
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
-    @patch("common.enrichment.requests.head")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network.requests.head")
     def test_get_fallback_ssrf_blocked(self, mock_head, _mock_b64, _mock_gnews):  # noqa: PT019
         # HEAD section: is_private_url returns False first (line 395) → HEAD proceeds
         # HEAD raises RequestException → caught at line 434 → falls to GET section
@@ -2971,16 +2971,16 @@ class TestResolveGoogleNewsInnerGetSSRF:
         mock_head.side_effect = req_mod.exceptions.ConnectionError("refused")
         # First call (line 395 in HEAD section): False → proceed
         # Second call (line 439 in GET section): True → SSRF block
-        with patch("common.enrichment.is_private_url", side_effect=[False, True]):
+        with patch("common.enrichment_network.is_private_url", side_effect=[False, True]):
             result = _enrich_mod._resolve_google_news_url_inner(
                 "https://news.google.com/rss/articles/CBMiGETSSRF", timeout=1
             )
         assert result == ""
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
-    @patch("common.enrichment.requests.head")
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network.requests.head")
+    @patch("common.enrichment_network.requests.get")
     def test_get_fallback_executed_after_head_exception(self, mock_get, mock_head, _mock_b64, _mock_gnews):  # noqa: PT019
         # HEAD throws → GET fallback is attempted; not SSRF-blocked → normal execution
         import requests as req_mod
@@ -2990,7 +2990,7 @@ class TestResolveGoogleNewsInnerGetSSRF:
         mock_get_resp.url = "https://real-news.com/article"
         mock_get_resp.text = ""
         mock_get.return_value = mock_get_resp
-        with patch("common.enrichment.is_private_url", return_value=False):
+        with patch("common.enrichment_network.is_private_url", return_value=False):
             result = _enrich_mod._resolve_google_news_url_inner(
                 "https://news.google.com/rss/articles/CBMiGETOK", timeout=1
             )
@@ -3007,10 +3007,10 @@ class TestResolveGoogleNewsInnerGetRespUrlSSRF:
     def setup_method(self):
         _enrich_mod._gnews_url_cache.clear()
 
-    @patch("common.enrichment._resolve_via_gnewsdecoder", return_value="")
-    @patch("common.enrichment._decode_google_news_base64", return_value="")
-    @patch("common.enrichment.requests.head")
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network._resolve_via_gnewsdecoder", return_value="")
+    @patch("common.enrichment_network._decode_google_news_base64", return_value="")
+    @patch("common.enrichment_network.requests.head")
+    @patch("common.enrichment_network.requests.get")
     def test_resp_url_private_returns_empty(self, mock_get, mock_head, _mock_b64, _mock_gnews):  # noqa: PT019
         import requests as req_mod
 
@@ -3019,7 +3019,7 @@ class TestResolveGoogleNewsInnerGetRespUrlSSRF:
         mock_get_resp.url = "http://172.16.0.1/internal"
         mock_get_resp.text = ""
         mock_get.return_value = mock_get_resp
-        with patch("common.enrichment.is_private_url", side_effect=lambda u: "172.16" in u):
+        with patch("common.enrichment_network.is_private_url", side_effect=lambda u: "172.16" in u):
             result = _enrich_mod._resolve_google_news_url_inner(
                 "https://news.google.com/rss/articles/CBMiGETRESP", timeout=1
             )
@@ -3048,7 +3048,7 @@ class TestResolveViaGnewsdecoder:
         }
         with (
             patch.dict("sys.modules", {"googlenewsdecoder": mock_gnews_mod}),
-            patch("common.enrichment.is_private_url", return_value=False),
+            patch("common.enrichment_network.is_private_url", return_value=False),
         ):
             result = _resolve_via_gnewsdecoder("https://news.google.com/rss/articles/CBMi")
         assert result == "https://real-article.com/news"
@@ -3062,7 +3062,7 @@ class TestResolveViaGnewsdecoder:
         }
         with (
             patch.dict("sys.modules", {"googlenewsdecoder": mock_gnews_mod}),
-            patch("common.enrichment.is_private_url", return_value=True),
+            patch("common.enrichment_network.is_private_url", return_value=True),
         ):
             result = _resolve_via_gnewsdecoder("https://news.google.com/rss/articles/CBMi")
         assert result == ""
@@ -3109,7 +3109,7 @@ class TestIsValidImageUrlLongGif:
 # ---------------------------------------------------------------------------
 class TestFetchOgImageSSRF:
     def test_private_url_returns_empty(self):
-        with patch("common.enrichment.is_private_url", return_value=True):
+        with patch("common.enrichment_network.is_private_url", return_value=True):
             result = _fetch_og_image("https://internal.corp/image")
         assert result == ""
 
@@ -3122,8 +3122,8 @@ class TestFetchOgImageSSRF:
 # Lines 560-561: _fetch_og_image — non-http og:image scheme rejected
 # ---------------------------------------------------------------------------
 class TestFetchOgImageNonHttpScheme:
-    @patch("common.enrichment.is_private_url", return_value=False)
-    @patch("common.enrichment.requests.get")
+    @patch("common.enrichment_network.is_private_url", return_value=False)
+    @patch("common.enrichment_network.requests.get")
     def test_relative_image_url_skipped_falls_through(self, mock_get, _mock_priv):  # noqa: PT019
         # og:image with relative URL → skipped → function returns ""
         html = (
